common: power: Weaken hispeed logic when screen is off
When the screen	is off, our workload is	non-
latency-sensitive, so we can set go_hispeed_load
to 100 while the screen is off.

The result is that the boost to hispeed_freq
and the connected possible boost to max_freq
happens less often.

This is added to the targets which are known
to support mpctlv3: msm8952, msm8992,
msm8994 and msm8996.

Change-Id: I6d1ce4c560bbc8aa86c5b022513f244dd8c780d8
Signed-off-by: Alex Naidis <alex.naidis@linux.com>